Added an inefficient optimizer
There is now an optional optimization pass that runs unit resolution
over the clauses to completion.  Unit resolution here is basically
constant folding.  So it will find all the propositions that are
constants and set them to their constant values.

At some point, I'll make the optimizer faster.  But it only needs to run
once problem definition time, not each time a solve is done.  So
efficiency should be less important.